# docs/api/openapi.v1.json
{
  "components": {
    "schemas": {
...
      "ApiResponse_ErrorDto": {
        "description": "Standard API response envelope.\n\nAll successful responses are wrapped in this structure to provide\nconsistent pagination metadata and a uniform client interface.",
        "properties": {
          "data": {
            "description": "Error response body.",
            "properties": {
              "error": {
                "type": "string"
              }
            },
            "required": [
              "error"
            ],
            "type": "object"
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta",
                "description": "Pagination metadata (present for list endpoints)."
              }
            ]
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "ApiResponse_PostDto": {
        "description": "Standard API response envelope.\n\nAll successful responses are wrapped in this structure to provide\nconsistent pagination metadata and a uniform client interface.",
        "properties": {
          "data": {
            "description": "A post with full metadata.\n\nUsed across all endpoints that return post data (search, recent,\nsimilar, single post). The `score` field is only populated for search\nand similarity results.",
            "properties": {
              "direct_image_url": {
                "description": "URL to the full-resolution image (falls back to thumbnail if unavailable).",
                "type": "string"
              },
              "height": {
                "description": "Image height in pixels (0 if unknown).",
                "format": "int32",
                "minimum": 0,
                "type": "integer"
              },
              "ingestion_date": {
                "description": "Unix timestamp of when the post was ingested.",
                "format": "int64",
                "type": "integer"
              },
              "post_id": {
                "description": "Site-local post ID.",
                "format": "int64",
                "minimum": 0,
                "type": "integer"
              },
              "post_url": {
                "description": "Canonical URL to the post page on the source site.",
                "type": "string"
              },
              "rating": {
                "description": "Content rating (e.g. `\"s\"`, `\"q\"`, `\"e\"`).",
                "type": [
                  "string",
                  "null"
                ]
              },
              "score": {
                "description": "Cosine similarity score in [0.0, 1.0]; only set for search/similar results.",
                "format": "float",
                "type": [
                  "number",
                  "null"
                ]
              },
              "site": {
                "description": "Site name this post belongs to (e.g. `\"rule34\"`, `\"e621\"`).",
                "type": "string"
              },
              "tags": {
                "description": "Space-separated tags describing the image content.",
                "type": [
                  "string",
                  "null"
                ]
              },
              "thumbnail_url": {
                "description": "URL to a small thumbnail image.",
                "type": "string"
              },
              "width": {
                "description": "Image width in pixels (0 if unknown).",
                "format": "int32",
                "minimum": 0,
                "type": "integer"
              }
            },
            "required": [
              "post_id",
              "site",
              "post_url",
              "thumbnail_url",
              "direct_image_url",
              "width",
              "height",
              "ingestion_date"
            ],
            "type": "object"
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta",
                "description": "Pagination metadata (present for list endpoints)."
              }
            ]
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "ApiResponse_Vec_PostDto": {
        "description": "Standard API response envelope.\n\nAll successful responses are wrapped in this structure to provide\nconsistent pagination metadata and a uniform client interface.",
        "properties": {
          "data": {
            "items": {
              "description": "A post with full metadata.\n\nUsed across all endpoints that return post data (search, recent,\nsimilar, single post). The `score` field is only populated for search\nand similarity results.",
              "properties": {
                "direct_image_url": {
                  "description": "URL to the full-resolution image (falls back to thumbnail if unavailable).",
                  "type": "string"
                },
                "height": {
                  "description": "Image height in pixels (0 if unknown).",
                  "format": "int32",
                  "minimum": 0,
                  "type": "integer"
                },
                "ingestion_date": {
                  "description": "Unix timestamp of when the post was ingested.",
                  "format": "int64",
                  "type": "integer"
                },
                "post_id": {
                  "description": "Site-local post ID.",
                  "format": "int64",
                  "minimum": 0,
                  "type": "integer"
                },
                "post_url": {
                  "description": "Canonical URL to the post page on the source site.",
                  "type": "string"
                },
                "rating": {
                  "description": "Content rating (e.g. `\"s\"`, `\"q\"`, `\"e\"`).",
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "score": {
                  "description": "Cosine similarity score in [0.0, 1.0]; only set for search/similar results.",
                  "format": "float",
                  "type": [
                    "number",
                    "null"
                  ]
                },
                "site": {
                  "description": "Site name this post belongs to (e.g. `\"rule34\"`, `\"e621\"`).",
                  "type": "string"
                },
                "tags": {
                  "description": "Space-separated tags describing the image content.",
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "thumbnail_url": {
                  "description": "URL to a small thumbnail image.",
                  "type": "string"
                },
                "width": {
                  "description": "Image width in pixels (0 if unknown).",
                  "format": "int32",
                  "minimum": 0,
                  "type": "integer"
                }
              },
              "required": [
                "post_id",
                "site",
                "post_url",
                "thumbnail_url",
                "direct_image_url",
                "width",
                "height",
                "ingestion_date"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta",
                "description": "Pagination metadata (present for list endpoints)."
              }
            ]
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "ResponseMeta": {
        "description": "Pagination and query metadata.",
        "properties": {
          "count": {
            "description": "Number of items returned in this response.",
            "minimum": 0,
            "type": "integer"
          },
          "next_offset": {
            "description": "Offset for the next page (None if no more results).",
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "count"
        ],
        "type": "object"
      },
...
      "SearchUploadForm": {
        "description": "Multipart form fields for `/api/search/upload`.",
        "properties": {
          "image": {
            "description": "Uploaded image file used as visual query.",
            "format": "binary",
            "type": [
              "string",
              "null"
            ]
          },
          "image_weight": {
            "description": "Image weight for hybrid text+image queries, in [0.0, 1.0].",
            "example": 0.5,
            "format": "float",
            "maximum": 1,
            "minimum": 0,
            "type": [
              "number",
              "null"
            ]
          },
          "limit": {
            "description": "Maximum number of results to return (default: 10, max: 100).",
            "example": 10,
            "format": "int64",
            "maximum": 100,
            "minimum": 1,
            "type": [
              "integer",
              "null"
            ]
          },
          "q": {
            "description": "Text query for semantic search.",
            "example": "fluffy cat sitting on a windowsill",
            "type": [
              "string",
              "null"
            ]
          },
          "site": {
            "description": "Sites to restrict results to. Comma-separated.",
            "example": "rule34,e621",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Semantic image search API for booru sites. Supports text queries, image URL queries, and hybrid text+image searches with optional per-site filtering.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "Roobu API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/post/{site}/{post_id}": {
      "get": {
        "operationId": "get_post",
        "parameters": [
          {
            "description": "Site name (e.g. rule34, e621)",
            "in": "path",
            "name": "site",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Site-local post ID",
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "format": "int64",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PostDto"
                }
              }
            },
            "description": "Post found"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ErrorDto"
                }
              }
            },
            "description": "Post not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ErrorDto"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Get a single post by site and ID.",
        "tags": [
          "posts"
        ]
      }
    },
...
    "/api/search": {
      "get": {
        "operationId": "search",
        "parameters": [
          {
            "description": "Text query for semantic search",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Publicly accessible image URL for visual search",
            "in": "query",
            "name": "image_url",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Sites to filter by (repeated or comma-separated)",
            "in": "query",
            "name": "site",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "description": "Maximum results (default 10, max 100)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int64",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Image weight for hybrid queries [0.0-1.0]",
            "in": "query",
            "name": "image_weight",
            "required": false,
            "schema": {
              "format": "float",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_PostDto"
                }
              }
            },
            "description": "Search results"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ErrorDto"
                }
              }
            },
            "description": "Bad request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ErrorDto"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Semantic search over indexed images.",
        "tags": [
          "search"
        ]
      }
    },
    "/api/search/similar/{site}/{post_id}": {
      "get": {
        "operationId": "search_similar",
        "parameters": [
          {
            "description": "Site name of the target post",
            "in": "path",
            "name": "site",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Site-local post ID of the target",
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "format": "int64",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Maximum results (default 10, max 100)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int64",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_PostDto"
                }
              }
            },
            "description": "Similar posts"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ErrorDto"
                }
              }
            },
            "description": "Target post not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ErrorDto"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Find visually similar images to a known post.",
        "tags": [
          "search"
        ]
      }
    },
    "/api/search/upload": {
      "post": {
        "operationId": "search_upload",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SearchUploadForm"
              }
            }
          },
          "description": "Multipart form with optional q text, optional image file, optional comma-separated site, limit, and image_weight",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_PostDto"
                }
              }
            },
            "description": "Search results"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ErrorDto"
                }
              }
            },
            "description": "Bad request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ErrorDto"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Semantic search using multipart form-data, including direct image upload.",
        "tags": [
          "search"
        ]
      }
    },
...
  },
...
}